requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
streamlit
requests
python-dotenv
SQLAlchemy
psycopg2-binary
reportlab
PyPDF2
